# docs/swagger.yaml
definitions:
  wallet.Err:
    properties:
      message:
        type: string
    type: object
  wallet.Wallet:
    properties:
      balance:
        example: 100
        type: number
      created_at:
        example: "2024-03-25T14:19:00.729237Z"
        type: string
      id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
      user_name:
        example: John Doe
        type: string
      wallet_name:
        example: John's Wallet
        type: string
      wallet_type:
        example: Create Card
        type: string
    type: object
host: localhost:1323
info:
  contact: {}
  description: Sophisticated Wallet API
  title: Wallet API
  version: "1.1"
paths:
  /api/v1/users/:user_id/wallets:
    delete:
      consumes:
      - application/json
      description: Delete a wallet by User Id
      parameters:
      - description: user_id
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Delete a wallet by User Id
      tags:
      - wallet
    get:
      consumes:
      - application/json
      description: Get a wallet by User Id
      parameters:
      - description: user_id
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Get a wallet by User Id
      tags:
      - wallet
  /api/v1/wallets:
    get:
      consumes:
      - application/json
      description: Get all wallets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Get all wallets
      tags:
      - wallet
    patch:
      consumes:
      - application/json
      description: Update a wallet
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Update a wallet
      tags:
      - wallet
    post:
      consumes:
      - application/json
      description: Create a wallet
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Create a wallet
      tags:
      - wallet
  /api/v1/wallets/:wallet-type:
    get:
      consumes:
      - application/json
      description: Get a wallet by Wallet type
      parameters:
      - description: wallet-type
        in: path
        name: wallet-type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Get a wallet by Wallet type
      tags:
      - wallet
swagger: "2.0"